+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -4,7 +4,7 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
-* Fixed dropdown not opening on pressing space in firefox.
+* Fixed dropdown not opening on pressing space in firefox, and reverted disabled tests for firefox.
 * Disabled wdio test which opens dropdown menu with space key for firefox testing. This will be re-enabled when https://github.com/cerner/terra-core/issues/3008 is resolved.
 
 ### Removed